# docs/swagger.yaml
basePath: /api/v1
definitions:
  api.Response:
    properties:
      message:
        type: string
      status_code:
        type: integer
    type: object
  controller.AuthFormInput:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  database.Body:
    properties:
      url:
        type: string
    type: object
  database.StatusMessage:
    properties:
      data:
        type: string
      message:
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a url shortener server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: URL Shortener API
  version: "1.0"
paths:
  /address/{url-id}:
    delete:
      consumes:
      - application/json
      description: deletes short urls using the url id
      parameters:
      - description: url id in database
        in: path
        name: url-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.StatusMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/database.StatusMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/database.StatusMessage'
      summary: delete address and short code from database
      tags:
      - url
  /address/{url}:
    get:
      consumes:
      - application/json
      description: returns the full url of the short code
      parameters:
      - description: shortcode to url
        in: path
        name: url
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.StatusMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/database.StatusMessage'
      summary: Full address of short code
      tags:
      - url
  /address/shorten:
    post:
      consumes:
      - application/json
      description: returns the short code of url shortened
      parameters:
      - description: url to shorten
        in: body
        name: url
        required: true
        schema:
          $ref: '#/definitions/database.Body'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/database.StatusMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/database.StatusMessage'
      summary: Shortens a url
      tags:
      - url
  /signin:
    post:
      consumes:
      - application/json
      description: login user and create user session
      parameters:
      - description: username and password
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/controller.AuthFormInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response'
      summary: signin user
      tags:
      - auth
  /signout:
    post:
      consumes:
      - application/json
      description: logout user and destroy user session
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response'
      summary: signout user
      tags:
      - auth
  /signup:
    post:
      consumes:
      - application/json
      description: register user details to database
      parameters:
      - description: username and password
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/controller.AuthFormInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response'
      summary: signup user
      tags:
      - auth
swagger: "2.0"